gluon-luci-admin: fix SSH keys being removed when a password is set

+++ package/gluon-luci-admin/files/usr/lib/lua/luci/model/cbi/admin/remote.lua
@@ -51,7 +51,9 @@ if fs.access("/etc/config/dropbear") then
   end
 
   function keys.remove(self, section)
-    fs.remove("/etc/dropbear/authorized_keys")
+    if keys:formvalue("_keys") then
+      fs.remove("/etc/dropbear/authorized_keys")
+    end
   end
 end
